Move YAML deserializer into util crate
