+++ Cargo.toml
@@ -42,6 +42,7 @@ mime = "0.3.17"
 pretty_assertions = "1.4.0"
 reqwest = {version = "0.12.5", default-features = false}
 rstest = {version = "0.24.0", default-features = false}
+saphyr = "0.0.6"
 schemars = "1.0.2"
 serde = {version = "1.0.204", default-features = false}
 serde_json = {version = "1.0.120", default-features = false, features = ["preserve_order"]}
