+++ crates/core/Cargo.toml
@@ -26,7 +26,7 @@ reqwest = {workspace = true, features = ["json", "multipart", "rustls-tls", "rus
 rstest = {workspace = true, optional = true}
 rusqlite = {version = "0.35.0", default-features = false, features = ["bundled", "chrono", "uuid"]}
 rusqlite_migration = "2.1.0"
-saphyr = "0.0.6"
+saphyr = {workspace = true}
 schemars = {workspace = true, features = ["indexmap2"], optional = true}
 serde = {workspace = true, features = ["derive"]}
 serde_json = {workspace = true}
