+++ crates/core/src/collection.rs
@@ -5,7 +5,6 @@ mod cereal;
 mod json;
 mod models;
 mod recipe_tree;
-mod resolve;
 #[cfg(feature = "schema")]
 mod schema;
