+++ crates/core/src/collection/cereal.rs
@@ -12,21 +12,20 @@ use crate::{
         Authentication, Collection, Folder, JsonTemplate, Profile, ProfileId,
         QueryParameterValue, Recipe, RecipeBody, RecipeId, RecipeTree,
         recipe_tree::RecipeNode,
-        resolve::{ReferenceError, ResolveReferences},
     },
     http::HttpMethod,
 };
 use indexmap::IndexMap;
-use itertools::Itertools;
-use saphyr::{
-    AnnotatedMapping, LoadableYamlNode, MarkedYaml, Marker, Scalar, ScanError,
-    YamlData,
-};
+use saphyr::{LoadableYamlNode, MarkedYaml, Scalar, YamlData};
 use slumber_template::Template;
+use slumber_util::{
+    deserialize_enum, impl_deserialize_from,
+    yaml::{
+        DeserializeYaml, Error, Expected, Field, LocatedError, MarkedYamlExt,
+        ResolveReferences, StructDeserializer, yaml_parse_panic,
+    },
+};
 use std::path::Path;
-use thiserror::Error;
-
-const TYPE_FIELD: &str = "type";
 
 type Result<T> = std::result::Result<T, LocatedError<Error>>;
 
@@ -83,73 +82,6 @@ pub fn deserialize_collection(
     )
 }
 
-/// Deserialize from YAML into the implementing type
-trait DeserializeYaml: Sized {
-    /// What kind of YAML value do we expect to see?
-    fn expected() -> Expected;
-
-    /// Deserialize the given YAML value into this type
-    fn deserialize(yaml: MarkedYaml) -> Result<Self>;
-}
-
-/// Implement [DeserializeYaml] for a type `T` via type `U`, where `T: From<U>,
-/// U: DeserializeYaml`
-macro_rules! impl_deserialize_from {
-    ($t:ty, $u:ty) => {
-        impl DeserializeYaml for $t {
-            fn expected() -> Expected {
-                <$u as DeserializeYaml>::expected()
-            }
-
-            fn deserialize(yaml: MarkedYaml) -> Result<Self> {
-                <$u as DeserializeYaml>::deserialize(yaml).map(<$t>::from)
-            }
-        }
-    };
-}
-
-/// Deserialize a YAML value as an internally tagged enum. The `type` field will
-/// contain the variant, and all other fields in the mapping will be
-/// deserialized using the given function.
-macro_rules! deserialize_enum {
-    ($yaml:expr, $($tag:literal => $f:expr),* $(,)?) => {
-            const EXPECTED: Expected =
-                Expected::OneOf(&[$(&Expected::Literal($tag),)*]);
-
-            // Find the enum variant based on the `type` field
-            let span = $yaml.span;
-            let mut mapping = $yaml.try_into_mapping()?;
-            let kind_yaml = mapping
-                .remove(&MarkedYaml::value_from_str(TYPE_FIELD))
-                .ok_or(LocatedError {
-                    error: Error::MissingField {
-                        field: TYPE_FIELD,
-                        expected: EXPECTED,
-                    },
-                    location: span.start,
-                })?;
-            let kind_location = kind_yaml.span.start;
-            let kind = kind_yaml.try_into_string()?;
-
-            // Deserialize the rest of the mapping as the specified enum variant
-            let yaml = MarkedYaml {
-                data: YamlData::Mapping(mapping),
-                span,
-            };
-            match kind.as_str() {
-                $($tag => $f(yaml),)*
-                // Unknown tag
-                _ => Err(LocatedError {
-                    error: Error::Unexpected {
-                        expected: EXPECTED,
-                        actual: format!("{kind:?}"),
-                    },
-                    location: kind_location,
-                }),
-            }
-    };
-}
-
 impl_deserialize_from!(ProfileId, String);
 impl_deserialize_from!(RecipeId, String);
 
@@ -168,7 +100,9 @@ impl DeserializeYaml for Collection {
 
         let collection = Collection {
             name: deserializer.get(Field::new("name").opt())?,
-            profiles: deserializer.get(Field::new("profiles").opt())?,
+            profiles: deserializer
+                .get::<Adopt<_>>(Field::new("profiles").opt())?
+                .0,
             // Internally we call these recipes, but extensive market research
             // shows that `requests` is more intuitive to the user
             recipes: deserializer.get(Field::new("requests").opt())?,
@@ -181,7 +115,7 @@ impl DeserializeYaml for Collection {
 /// Deserialize a map of profiles. This needs a custom implementation because:
 /// - To call [HasId::set_id] on each value
 /// - We have to enforce that at most one profile is set as default
-impl DeserializeYaml for IndexMap<ProfileId, Profile> {
+impl DeserializeYaml for Adopt<IndexMap<ProfileId, Profile>> {
     fn expected() -> Expected {
         Expected::Mapping
     }
@@ -201,13 +135,13 @@ impl DeserializeYaml for IndexMap<ProfileId, Profile> {
                 // Check if another profile is already the default
                 if value.default {
                     if let Some(default) = default_profile.take() {
-                        return Err(LocatedError {
-                            error: Error::MultipleDefaultProfiles {
+                        return Err(LocatedError::other(
+                            CerealError::MultipleDefaultProfiles {
                                 first: default,
                                 second: key,
                             },
-                            location: value_location,
-                        });
+                            value_location,
+                        ));
                     }
 
                     default_profile = Some(key.clone());
@@ -215,7 +149,8 @@ impl DeserializeYaml for IndexMap<ProfileId, Profile> {
 
                 Ok((key, value))
             })
-            .collect()
+            .collect::<Result<_>>()
+            .map(Adopt)
     }
 }
 
@@ -244,9 +179,10 @@ impl DeserializeYaml for RecipeTree {
 
     fn deserialize(yaml: MarkedYaml) -> Result<Self> {
         let location = yaml.span.start;
-        let recipes = IndexMap::deserialize(yaml)?;
+        let recipes: Adopt<IndexMap<RecipeId, RecipeNode>> =
+            Adopt::deserialize(yaml)?;
         // Build a tree from the map
-        RecipeTree::new(recipes)
+        RecipeTree::new(recipes.0)
             .map_err(|error| LocatedError::other(error, location))
     }
 }
@@ -254,7 +190,7 @@ impl DeserializeYaml for RecipeTree {
 /// Deserialize a map of profiles. This needs a custom implementation to call
 /// [HasId::set_id] on each value. This is used for both the root `requests`
 /// field and each inner folder.
-impl DeserializeYaml for IndexMap<RecipeId, RecipeNode> {
+impl DeserializeYaml for Adopt<IndexMap<RecipeId, RecipeNode>> {
     fn expected() -> Expected {
         Expected::Mapping
     }
@@ -268,7 +204,8 @@ impl DeserializeYaml for IndexMap<RecipeId, RecipeNode> {
                 value.set_id(key.clone());
                 Ok((key, value))
             })
-            .collect()
+            .collect::<Result<_>>()
+            .map(Adopt)
     }
 }
 
@@ -297,10 +234,10 @@ impl DeserializeYaml for RecipeNode {
         } else if has("requests") {
             Folder::deserialize(yaml).map(RecipeNode::Folder)
         } else {
-            Err(LocatedError {
-                error: Error::UnknownRecipeNodeVariant,
-                location: yaml.span.start,
-            })
+            Err(LocatedError::other(
+                CerealError::UnknownRecipeNodeVariant,
+                yaml.span.start,
+            ))
         }
     }
 }
@@ -325,7 +262,7 @@ impl DeserializeYaml for Recipe {
             // Lower-case all headers for consistency. HTTP/1.1 headers are
             // case-insensitive and HTTP/2 enforces lower casing.
             headers: deserializer
-                .get(Field::<IndexMap<String, Template>>::new("headers").opt())?
+                .get::<IndexMap<String, Template>>(Field::new("headers").opt())?
                 .into_iter()
                 .map(|(k, v)| (k.to_lowercase(), v))
                 .collect(),
@@ -346,7 +283,9 @@ impl DeserializeYaml for Folder {
             id: RecipeId::default(), // Will be set by parent based on key
             name: deserializer.get(Field::new("name").opt())?,
             // `requests` matches the root field name
-            children: deserializer.get(Field::new("requests").opt())?,
+            children: deserializer
+                .get::<Adopt<_>>(Field::new("requests").opt())?
+                .0,
         };
         deserializer.done()?;
         Ok(folder)
@@ -465,14 +404,14 @@ impl DeserializeYaml for JsonTemplate {
             YamlData::Value(Scalar::Null) => Ok(Self::Null),
             YamlData::Value(Scalar::Boolean(b)) => Ok(Self::Bool(b)),
             YamlData::Value(Scalar::Integer(i)) => Ok(Self::Number(i.into())),
-            YamlData::Value(Scalar::FloatingPoint(f)) => {
-                Ok(Self::Number(serde_json::Number::from_f64(f.0).ok_or_else(
-                    || LocatedError {
-                        error: Error::InvalidJsonFloat(f.0),
-                        location: yaml.span.start,
-                    },
-                )?))
-            }
+            YamlData::Value(Scalar::FloatingPoint(f)) => Ok(Self::Number(
+                serde_json::Number::from_f64(f.0).ok_or_else(|| {
+                    LocatedError::other(
+                        CerealError::InvalidJsonFloat(f.0),
+                        yaml.span.start,
+                    )
+                })?,
+            )),
             // Parse string as a template
             YamlData::Value(Scalar::String(s)) => {
                 let template = s.parse::<Template>().map_err(|error| {
@@ -505,258 +444,19 @@ impl DeserializeYaml for JsonTemplate {
     }
 }
 
-impl DeserializeYaml for Template {
-    fn expected() -> Expected {
-        Expected::OneOf(&[
-            &Expected::String,
-            &Expected::Boolean,
-            &Expected::Number,
-            // We accept `null` too, but it's not a helpful suggestion
-        ])
-    }
-
-    fn deserialize(yaml: MarkedYaml) -> Result<Self> {
-        if let YamlData::Value(scalar) = yaml.data {
-            // Accept any scalar for a template. We'll treat everything as the
-            // equivalent string representation
-            match scalar {
-                Scalar::Null => "null".parse(),
-                Scalar::Boolean(b) => b.to_string().parse(),
-                Scalar::Integer(i) => i.to_string().parse(),
-                Scalar::FloatingPoint(f) => f.to_string().parse(),
-                Scalar::String(s) => s.parse(),
-            }
-            .map_err(|error| LocatedError::other(error, yaml.span.start))
-        } else {
-            Err(LocatedError::unexpected(Expected::String, yaml))
-        }
-    }
-}
-
-impl DeserializeYaml for bool {
-    fn expected() -> Expected {
-        Expected::Boolean
-    }
-
-    fn deserialize(yaml: MarkedYaml) -> Result<Self> {
-        yaml.try_into_bool()
-    }
-}
-
-impl DeserializeYaml for String {
-    fn expected() -> Expected {
-        Expected::String
-    }
-
-    fn deserialize(yaml: MarkedYaml) -> Result<Self> {
-        yaml.try_into_string()
-    }
-}
-
-impl<T: DeserializeYaml> DeserializeYaml for Option<T> {
-    fn expected() -> Expected {
-        // Techinically we should include `null` here too, but generally
-        // optional fields should just be omitted instead of being set to null.
-        // It also makes the lifetimes and type signatures on Expected much more
-        // complicated to dynamically build one that's not 'static
-        T::expected()
-    }
-
-    fn deserialize(yaml: MarkedYaml) -> Result<Self> {
-        if yaml.data.is_null() {
-            Ok(None)
-        } else {
-            T::deserialize(yaml).map(Some)
-        }
-    }
-}
-
-impl<T> DeserializeYaml for Vec<T>
-where
-    T: DeserializeYaml,
-{
-    fn expected() -> Expected {
-        Expected::Sequence
-    }
-
-    fn deserialize(yaml: MarkedYaml) -> Result<Self> {
-        let sequence = yaml.try_into_sequence()?;
-        sequence.into_iter().map(T::deserialize).collect()
-    }
-}
-
-/// Deserialize a plain map with string keys
-impl<V> DeserializeYaml for IndexMap<String, V>
-where
-    V: DeserializeYaml,
-{
-    fn expected() -> Expected {
-        Expected::Mapping
-    }
-
-    fn deserialize(yaml: MarkedYaml) -> Result<Self> {
-        yaml.try_into_mapping()?
-            .into_iter()
-            .map(|(k, v)| Ok((k.try_into_string()?, V::deserialize(v)?)))
-            .collect()
-    }
-}
-
-/// Extension trait to add fallible conversion methods to [MarkedYaml]
-trait MarkedYamlExt<'a>: Sized {
-    /// Unpack the YAML as a boolean
-    fn try_into_bool(self) -> Result<bool>;
-
-    /// Unpack the YAML as a string
-    fn try_into_string(self) -> Result<String>;
-
-    /// Unpack the YAML as a sequence
-    fn try_into_sequence(self) -> Result<Vec<Self>>;
-
-    /// Unpack the YAML as a mapping
-    fn try_into_mapping(self) -> Result<AnnotatedMapping<'a, Self>>;
-}
-
-impl<'a> MarkedYamlExt<'a> for MarkedYaml<'a> {
-    fn try_into_bool(self) -> Result<bool> {
-        if let YamlData::Value(Scalar::Boolean(b)) = self.data {
-            Ok(b)
-        } else {
-            Err(LocatedError::unexpected(Expected::Boolean, self))
-        }
-    }
-
-    fn try_into_string(self) -> Result<String> {
-        if let YamlData::Value(Scalar::String(s)) = self.data {
-            Ok(s.into_owned())
-        } else {
-            Err(LocatedError::unexpected(Expected::String, self))
-        }
-    }
-
-    fn try_into_sequence(self) -> Result<Vec<Self>> {
-        if let YamlData::Sequence(sequence) = self.data {
-            Ok(sequence)
-        } else {
-            Err(LocatedError::unexpected(Expected::Sequence, self))
-        }
-    }
-
-    fn try_into_mapping(self) -> Result<AnnotatedMapping<'a, Self>> {
-        if let YamlData::Mapping(mapping) = self.data {
-            Ok(mapping)
-        } else {
-            Err(LocatedError::unexpected(Expected::Mapping, self))
-        }
-    }
-}
-
-/// An error paired with the source location in YAML where the error occurred
-///
-/// This type is internal to this module because there's no external application
-/// for it beyond stringification.
-#[derive(Debug, derive_more::Display)]
-#[display("{error}")]
-pub struct LocatedError<E> {
-    /// Error that occurred
-    pub error: E,
-    /// Location of the error within the YAML file
-    pub location: Marker,
-}
-
-impl LocatedError<Error> {
-    /// Create a new [Other](Self::Other) from any error type
-    fn other(
-        error: impl Into<Box<dyn std::error::Error + Send + Sync>>,
-        location: Marker,
-    ) -> Self {
-        Self {
-            error: Error::Other(error.into()),
-            location,
-        }
-    }
-
-    /// Create a new [UnexpectedType](Self::UnexpectedType) from the expected
-    /// type and actual value
-    fn unexpected(expected: Expected, actual: MarkedYaml) -> Self {
-        // Find a useful representation of the received value
-        let actual_string = match actual.data {
-            // Scalars are unlikely to be big so we can include the actual value
-            YamlData::Value(Scalar::Null) => "null".into(),
-            YamlData::Value(Scalar::Boolean(b)) => format!("`{b}`"),
-            YamlData::Value(Scalar::Integer(i)) => format!("`{i}`"),
-            YamlData::Value(Scalar::FloatingPoint(f)) => format!("`{f}`"),
-            // Use debug format to get wrapping quotes
-            YamlData::Value(Scalar::String(s)) => format!("{s:?}"),
-            YamlData::Tagged(tag, _) => format!("tag `{tag}`"),
-            // Collections could be large so just include the type
-            YamlData::Sequence(_) => "sequence".into(),
-            YamlData::Mapping(_) => "mapping".into(),
-            YamlData::Representation(_, _, _)
-            | YamlData::Alias(_)
-            | YamlData::BadValue => yaml_parse_panic(),
-        };
-        Self {
-            location: actual.span.start,
-            error: Error::Unexpected {
-                expected,
-                actual: actual_string,
-            },
-        }
-    }
-}
-
-impl From<ScanError> for LocatedError<Error> {
-    fn from(error: ScanError) -> Self {
-        Self {
-            location: *error.marker(),
-            error: Error::Scan(error),
-        }
-    }
-}
-
-/// An error that can occur while deserializing a YAML value
-#[derive(Debug, Error)]
-enum Error {
+/// A Slumber-specific error that can occur while deserializing a YAML value.
+/// Generic YAML errors are defined in [slumber_util::yaml::Error]. This only
+/// holds errors specific to collection deserialization.
+#[derive(Debug, thiserror::Error)]
+enum CerealError {
     /// JSON body contained a float value that isn't representable in JSON
     #[error("Invalid float `{0}`; JSON does not support NaN or Infinity")]
     InvalidJsonFloat(f64),
-
-    #[error("Expected field `{field}` with {expected}")]
-    MissingField {
-        field: &'static str,
-        expected: Expected,
-    },
 
     #[error(
         "Cannot set profile `{second}` as default; `{first}` is already default"
     )]
     MultipleDefaultProfiles { first: ProfileId, second: ProfileId },
-
-    /// External error type
-    #[error(transparent)]
-    Other(Box<dyn 'static + std::error::Error + Send + Sync>),
-
-    /// Error parsing or resolving a reference under a `$ref` tag
-    #[error(transparent)]
-    Reference(ReferenceError),
-
-    /// Error parsing YAML
-    #[error(transparent)]
-    Scan(saphyr::ScanError),
-
-    /// Expected a particular type or value, but received something else
-    #[error("Expected {expected}, received {actual}")]
-    Unexpected {
-        expected: Expected,
-        /// Pre-formatted "actual" value. Getting an owned YAML value from
-        /// is complicated so it's easier to store it as the presentation
-        /// string
-        actual: String,
-    },
-
-    #[error("Unexpected field `{0}`")]
-    UnexpectedField(String),
 
     /// We couldn't guess the variant of a recipe node based on its fields
     #[error(
@@ -764,125 +464,6 @@ enum Error {
         folders must have a `requests` field"
     )]
     UnknownRecipeNodeVariant,
-}
-
-/// When a value is expected but is either incorrect or missing, this type
-/// allows the caller to declare what they expected to find
-#[derive(Debug, derive_more::Display)]
-enum Expected {
-    /// Expected null
-    #[display("null")]
-    Null,
-    /// Expected a string
-    #[display("string")]
-    String,
-    /// Expected a boolean
-    #[display("boolean")]
-    Boolean,
-    /// Expected a number
-    #[display("number")]
-    Number,
-    /// Expected a sequence
-    #[display("sequence")]
-    Sequence,
-    /// Expected a mapping
-    #[display("mapping")]
-    Mapping,
-    /// Expected a string literal
-    #[display("{_0:?}")]
-    Literal(&'static str),
-    /// Expected one of a static set of strings (for enum discriminants)
-    #[display("one of {}", _0.iter().format(", "))]
-    OneOf(&'static [&'static Self]),
-}
-
-/// Utility for deserializing a struct or enum variant from a YAML mapping.
-/// Initialize this struct with a YAML value, and it will:
-/// - Ensure the value is a mapping
-/// - Enable deserializing individual fields with [get](Self::get)
-/// - Ensure no unexpected fields were present with [done](Self::done)
-///     - NOTE: `done` needs to be called manually after deserialization!
-struct StructDeserializer<'a> {
-    mapping: AnnotatedMapping<'a, MarkedYaml<'a>>,
-    location: Marker,
-}
-
-impl<'a> StructDeserializer<'a> {
-    fn new(yaml: MarkedYaml<'a>) -> Result<Self> {
-        let location = yaml.span.start;
-        let mapping = yaml.try_into_mapping()?;
-        Ok(Self { mapping, location })
-    }
-
-    /// Deserialize a field from the mapping
-    fn get<T: DeserializeYaml>(&mut self, field: Field<T>) -> Result<T> {
-        if let Some(value) =
-            self.mapping.remove(&MarkedYaml::value_from_str(field.name))
-        {
-            T::deserialize(value)
-        } else if let Some(default) = field.default {
-            Ok(default)
-        } else {
-            Err(LocatedError {
-                error: Error::MissingField {
-                    field: field.name,
-                    expected: T::expected(),
-                },
-                location: self.location,
-            })
-        }
-    }
-
-    /// Check that no fields were unused
-    fn done(mut self) -> Result<()> {
-        if let Some((key, _)) = self.mapping.pop_front() {
-            let key_location = key.span.start;
-            // If the key isn't a string, it's reasonable to return a type error
-            let key = key.try_into_string()?;
-            Err(LocatedError {
-                error: Error::UnexpectedField(key),
-                location: key_location,
-            })
-        } else {
-            Ok(())
-        }
-    }
-}
-
-/// A single deserializable field in a struct or enum variant. The field has a
-/// static name, which corresponds to the name of the field *in the YAML*.
-/// Generally this matches the internal field name, but not always. Fields are
-/// required by default, but can be made optional with [opt](Self::opt) or
-/// [or](Self::or).
-struct Field<T> {
-    name: &'static str,
-    default: Option<T>,
-}
-
-impl<T> Field<T> {
-    fn new(name: &'static str) -> Self {
-        Self {
-            name,
-            default: None,
-        }
-    }
-
-    /// Pre-populate this field with `T`'s default value. If the field is not
-    /// deserialized, the default value will be used instead.
-    fn opt(mut self) -> Self
-    where
-        T: Default,
-    {
-        self.default = Some(T::default());
-        self
-    }
-
-    /// Pre-populate this field with the given default value. If the field is
-    /// not deserialized, the default value will be used instead.
-    fn or(mut self, value: T) -> Self {
-        self.default = Some(value);
-        self
-    }
 }
 
 /// A type that has an `id` field. This is ripe for a derive macro, maybe a fun
@@ -937,14 +518,9 @@ impl HasId for Recipe {
     }
 }
 
-/// There are a few variants of [YamlData] that are not possible to encounter
-/// with the way we use the parser. They represent partially parsed data, while
-/// we do full parsing before starting deserialization. Call this function in
-/// `match` statements for these variants
-#[track_caller]
-pub fn yaml_parse_panic() -> ! {
-    unreachable!("Invalid or incomplete YAML data")
-}
+/// Workaround for the orphan rule
+#[derive(Debug, Default)]
+struct Adopt<T>(T);
 
 /// Expose this for RecipeTree's tests
 #[cfg(test)]
@@ -958,20 +534,22 @@ mod tests {
     use rstest::rstest;
     use serde_json::json;
     use serde_yaml::Mapping;
-    use slumber_util::assert_err;
-    use std::iter;
+    use slumber_util::{
+        assert_err,
+        yaml::{deserialize_yaml, yaml_enum, yaml_mapping},
+    };
 
     /// Test error cases for deserializing a profile map
     #[rstest]
     #[case::multiple_default(
-        mapping([
-            ("profile1", mapping([
+        yaml_mapping([
+            ("profile1", yaml_mapping([
                 ("default", serde_yaml::Value::Bool(true)),
-                ("data", mapping([("a", "1")]))
+                ("data", yaml_mapping([("a", "1")]))
             ])),
-            ("profile2", mapping([
+            ("profile2", yaml_mapping([
                 ("default", serde_yaml::Value::Bool(true)),
-                ("data", mapping([("a", "2")]))
+                ("data", yaml_mapping([("a", "2")]))
             ])),
         ]),
         "Cannot set profile `profile2` as default; `profile1` is already default",
@@ -981,7 +559,9 @@ mod tests {
         #[case] expected_error: &str,
     ) {
         assert_err!(
-            deserialize_yaml::<IndexMap<ProfileId, Profile>>(yaml.into()),
+            deserialize_yaml::<Adopt<IndexMap<ProfileId, Profile>>>(
+                yaml.into()
+            ),
             expected_error
         );
     }
@@ -996,7 +576,7 @@ mod tests {
     )]
     #[case::json(
         RecipeBody::json(json!({"user": "{{ user_id }}"})).unwrap(),
-        yaml_enum("json", [("data", mapping([("user", "{{ user_id }}")]))]),
+        yaml_enum("json", [("data", yaml_mapping([("user", "{{ user_id }}")]))]),
     )]
     #[case::json_nested(
         RecipeBody::json(json!(r#"{"warning": "NOT an object"}"#)).unwrap(),
@@ -1007,7 +587,7 @@ mod tests {
             "username".into() => "{{ username }}".into(),
             "password".into() => "{{ prompt('Password', sensitive=true) }}".into(),
         }),
-        yaml_enum("form_urlencoded", [("data", mapping([
+        yaml_enum("form_urlencoded", [("data", yaml_mapping([
             ("username", "{{ username }}"),
             ("password", "{{ prompt('Password', sensitive=true) }}"),
         ]))]),
@@ -1071,47 +651,10 @@ mod tests {
         );
     }
 
-    /// Deserialize a [serde_yaml::Value] using saphyr. Serde values are easier
-    /// to construct than saphyr values
-    fn deserialize_yaml<T: DeserializeYaml>(
-        yaml: serde_yaml::Value,
-    ) -> Result<T> {
-        let yaml_input = serde_yaml::to_string(&yaml).unwrap();
-        let mut documents = MarkedYaml::load_from_str(&yaml_input)?;
-        let yaml = documents.pop().unwrap();
-        T::deserialize(yaml)
-    }
-
     /// Helper for deserializing in RecipeTree's tests. We export this
     pub fn deserialize_recipe_tree(
         yaml: serde_yaml::Value,
     ) -> anyhow::Result<RecipeTree> {
         deserialize_yaml(yaml).map_err(|error| error.error.into())
-    }
-
-    /// Build a YAML mapping
-    fn mapping(
-        fields: impl IntoIterator<
-            Item = (&'static str, impl Into<serde_yaml::Value>),
-        >,
-    ) -> serde_yaml::Value {
-        fields
-            .into_iter()
-            .map(|(k, v)| (serde_yaml::Value::from(k), v.into()))
-            .collect::<Mapping>()
-            .into()
-    }
-
-    /// Build a YAML mapping with a `type` field
-    fn yaml_enum(
-        type_: &'static str,
-        fields: impl IntoIterator<
-            Item = (&'static str, impl Into<serde_yaml::Value>),
-        >,
-    ) -> serde_yaml::Value {
-        mapping(
-            iter::once((TYPE_FIELD, serde_yaml::Value::from(type_)))
-                .chain(fields.into_iter().map(|(k, v)| (k, v.into()))),
-        )
     }
 }
