+++ crates/template/src/cereal.rs
@@ -1,11 +1,13 @@
 use crate::{RenderError, Template, Value};
+use saphyr::{MarkedYaml, Scalar, YamlData};
 use serde::{
-    Deserialize, Deserializer, Serialize,
+    Serialize,
     de::{
-        self, Error, IntoDeserializer, Unexpected, Visitor,
+        self, IntoDeserializer, Unexpected, Visitor,
         value::{MapDeserializer, SeqDeserializer},
     },
 };
+use slumber_util::yaml::{DeserializeYaml, Expected, LocatedError};
 
 impl Serialize for Template {
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
@@ -16,55 +18,34 @@ impl Serialize for Template {
     }
 }
 
-// Custom deserializer for `Template`. This is useful for deserializing values
-// that are not strings, but should be treated as strings such as numbers,
-// booleans, and nulls.
-impl<'de> Deserialize<'de> for Template {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        struct TemplateVisitor;
+/// Deserialize templates via saphyr
+impl DeserializeYaml for Template {
+    fn expected() -> Expected {
+        Expected::OneOf(&[
+            &Expected::String,
+            &Expected::Boolean,
+            &Expected::Number,
+            // We accept `null` too, but it's not a helpful suggestion
+        ])
+    }
 
-        macro_rules! visit_primitive {
-            ($func:ident, $type:ty) => {
-                fn $func<E>(self, v: $type) -> Result<Self::Value, E>
-                where
-                    E: Error,
-                {
-                    self.visit_string(v.to_string())
-                }
-            };
-        }
-
-        impl Visitor<'_> for TemplateVisitor {
-            type Value = Template;
-
-            fn expecting(
-                &self,
-                formatter: &mut std::fmt::Formatter,
-            ) -> std::fmt::Result {
-                formatter.write_str("string, number, or boolean")
+    fn deserialize(
+        yaml: MarkedYaml,
+    ) -> Result<Self, LocatedError<slumber_util::yaml::Error>> {
+        if let YamlData::Value(scalar) = yaml.data {
+            // Accept any scalar for a template. We'll treat everything as the
+            // equivalent string representation
+            match scalar {
+                Scalar::Null => "null".parse(),
+                Scalar::Boolean(b) => b.to_string().parse(),
+                Scalar::Integer(i) => i.to_string().parse(),
+                Scalar::FloatingPoint(f) => f.to_string().parse(),
+                Scalar::String(s) => s.parse(),
             }
-
-            visit_primitive!(visit_bool, bool);
-            visit_primitive!(visit_u64, u64);
-            visit_primitive!(visit_i64, i64);
-            visit_primitive!(visit_f64, f64);
-
-            fn visit_str<E>(self, v: &str) -> Result<Self::Value, E>
-            where
-                E: Error,
-            {
-                // Minor hack: this error is always displayed followed the serde
-                // key path. Our error will contain a pointer to the error
-                // within the source text. For the pointer to line up, we have
-                // to start on a fresh line.
-                v.parse().map_err(|error| E::custom(format!("\n{error}")))
-            }
+            .map_err(|error| LocatedError::other(error, yaml.span.start))
+        } else {
+            Err(LocatedError::unexpected(Expected::String, yaml))
         }
-
-        deserializer.deserialize_any(TemplateVisitor)
     }
 }
 
@@ -177,25 +158,31 @@ impl<'de> serde::Deserializer<'de> for ValueDeserializer {
 mod tests {
     use super::*;
     use rstest::rstest;
-    use serde_test::{Token, assert_de_tokens};
+    use slumber_util::yaml::deserialize_yaml;
 
     /// Test deserialization, which has some additional logic on top of parsing
     #[rstest]
     // boolean
-    #[case::bool_true(Token::Bool(true), "true")]
-    #[case::bool_false(Token::Bool(false), "false")]
+    #[case::bool_true(true.into(), "true")]
+    #[case::bool_false(false.into(), "false")]
     // numeric
-    #[case::u64(Token::U64(1000), "1000")]
-    #[case::i64_negative(Token::I64(-1000), "-1000")]
-    #[case::float_positive(Token::F64(10.1), "10.1")]
-    #[case::float_negative(Token::F64(-10.1), "-10.1")]
+    #[case::u64(1000u64.into(), "1000")]
+    #[case::i64_negative((-1000i64).into(), "-1000")]
+    #[case::float_positive(10.1.into(), "10.1")]
+    #[case::float_negative((-10.1).into(), "-10.1")]
     // string
-    #[case::str(Token::Str("hello"), "hello")]
-    #[case::str_null(Token::Str("null"), "null")]
-    #[case::str_true(Token::Str("true"), "true")]
-    #[case::str_false(Token::Str("false"), "false")]
-    #[case::str_with_keys(Token::Str("{{ user_id }}"), "{{ user_id }}")]
-    fn test_deserialize_template(#[case] token: Token, #[case] expected: &str) {
-        assert_de_tokens(&Template::from(expected), &[token]);
+    #[case::str("hello".into(), "hello")]
+    #[case::str_null("null".into(), "null")]
+    #[case::str_true("true".into(), "true")]
+    #[case::str_false("false".into(), "false")]
+    #[case::str_with_keys("{{ user_id }}".into(), "{{ user_id }}")]
+    fn test_deserialize_template(
+        #[case] value: serde_yaml::Value,
+        #[case] expected: &str,
+    ) {
+        assert_eq!(
+            deserialize_yaml::<Template>(value).unwrap(),
+            Template::from(expected)
+        );
     }
 }
