+++ crates/util/Cargo.toml
@@ -15,12 +15,17 @@ tag = false
 
 [dependencies]
 anyhow = {workspace = true}
+derive_more = {workspace = true, features = ["display", "from"]}
 dirs = {workspace = true}
+indexmap = {workspace = true}
 itertools = {workspace = true}
+pretty_assertions = {workspace = true}
 rstest = {workspace = true, optional = true}
+saphyr = {workspace = true}
 serde = {workspace = true}
 serde_path_to_error = "0.1.16"
 serde_yaml = {workspace = true}
+thiserror = {workspace = true}
 tracing = {workspace = true}
 uuid = {workspace = true, features = ["v4"], optional = true}
 winnow = {workspace = true}
